Modified Kirby to support a different approach to multi-language
Not too many changes.  There is a new setting in config that hides the
default language. There is a whole section of code in sites.php that
seems to break multi-language support and causes Kirby to rewrite to a
parent folder instead of show a 404 error, so this has been commented
out.